% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load_disposable}
\alias{load_disposable}
\alias{load_disposable_}
\title{Create, install and load multiple disposable packages}
\usage{
load_disposable(..., lib_dir = tempfile(), imports = character())

load_disposable_(exprs, lib_dir = tempfile(), imports = character())
}
\arguments{
\item{...}{Named expressions.
A separate package with the given name is created for each.}

\item{lib_dir}{Directory to install the package to.
Defaults to a temporary directory that is
deleted once the R session is over.}

\item{imports}{The 'Imports' field in the DESCRIPTION file,
  the packages to import in each disposable package.}

\item{exprs}{Named quoted expressions to put in the packages.}
}
\description{
Create, install and load multiple disposable packages
}

