% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load_disposable}
\alias{load_disposable}
\title{Create, install and load multiple disposable packages}
\usage{
load_disposable(..., lib_dir = tempfile(), imports = character())
}
\arguments{
\item{...}{Named expressions.
A separate package with the given name is created for each.}

\item{lib_dir}{Directory to install the package to.
Defaults to a temporary directory that is
deleted once the R session is over.}

\item{imports}{The 'Imports' field in the DESCRIPTION file,
the packages to import in each disposable package.}
}
\value{
The \code{lib_dir}, invisibly.
}
\description{
Create, install and load multiple disposable packages
}
\examples{
\donttest{
lib_dir <- load_disposable(
  foo1 = { f <- function() print("hello!") ; d <- 1:10 },
  foo2 = { f <- function() print("hello again!") ; d <- 11:20 }
)
foo1::f()
foo2::f()
foo1::d
foo2::d
unloadNamespace("foo1")
unloadNamespace("foo2")
unlink(lib_dir, recursive = TRUE)
}
}

